Move Frontend.Page.Category to Frontend.Fragment.Category
