+++ src/Frontend/Page.hs
@@ -4,7 +4,6 @@ module Frontend.Page (module P)
 where
 
 import Frontend.Page.Admin      as P
-import Frontend.Page.Category   as P
 import Frontend.Page.Delegation as P
 import Frontend.Page.Idea       as P
 import Frontend.Page.Login      as P
