+++ src/Frontend/Page/Topic.hs
@@ -22,7 +22,7 @@ where
 
 import Action (ActionM, ActionPersist(..), ActionUserHandler, getCurrentTimestamp)
 import Control.Exception (assert)
-import Frontend.Page.Category
+import Frontend.Fragment.Category
 import Frontend.Page.Overview
 import Frontend.Prelude hiding (moveIdeasToLocation, editTopic)
