+++ tests/AllModulesSpec.hs
@@ -19,10 +19,10 @@ import DemoData ()
 import EventLog ()
 import Frontend ()
 import Frontend.Core ()
+import Frontend.Fragment.Category ()
 import Frontend.Fragment.Comment ()
 import Frontend.Page ()
 import Frontend.Page.Admin ()
-import Frontend.Page.Category ()
 import Frontend.Page.Delegation ()
 import Frontend.Page.Idea ()
 import Frontend.Page.Login ()
